Refactor pipe generation
Moved pipe generation to its own class
