+++ SE-CW2-UnityProject/Assets/Scenes/SampleScene.unity
@@ -1772,7 +1772,7 @@ Transform:
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children: []
   m_Father: {fileID: 0}
-  m_RootOrder: 9
+  m_RootOrder: 10
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1 &1440715874
 GameObject:
@@ -2147,7 +2147,7 @@ Transform:
   - {fileID: 1010201377}
   - {fileID: 1299691102}
   m_Father: {fileID: 0}
-  m_RootOrder: 10
+  m_RootOrder: 11
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!33 &1687824824
 MeshFilter:
@@ -2354,7 +2354,7 @@ Transform:
   - {fileID: 1131196264}
   - {fileID: 5385302}
   m_Father: {fileID: 0}
-  m_RootOrder: 11
+  m_RootOrder: 12
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!136 &1855687617
 CapsuleCollider:
@@ -2494,13 +2494,59 @@ RectTransform:
   - {fileID: 1998375546}
   - {fileID: 383918333}
   m_Father: {fileID: 0}
-  m_RootOrder: 8
+  m_RootOrder: 9
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0, y: 0}
   m_AnchorMax: {x: 0, y: 0}
   m_AnchoredPosition: {x: 0, y: 0}
   m_SizeDelta: {x: 0, y: 0}
   m_Pivot: {x: 0, y: 0}
+--- !u!1 &1903507867
+GameObject:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  serializedVersion: 6
+  m_Component:
+  - component: {fileID: 1903507869}
+  - component: {fileID: 1903507868}
+  m_Layer: 0
+  m_Name: pipe_Generation_object
+  m_TagString: Untagged
+  m_Icon: {fileID: 0}
+  m_NavMeshLayer: 0
+  m_StaticEditorFlags: 0
+  m_IsActive: 1
+--- !u!114 &1903507868
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1903507867}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 94f61a1eb0005a346bfebe434be8b430, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  small_Straight_Pipe: {fileID: 1687824822}
+  small_90Deg_Turn_Pipe: {fileID: 1855687612}
+  S_Pipe: {fileID: 2027925504}
+--- !u!4 &1903507869
+Transform:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1903507867}
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
+  m_LocalPosition: {x: 11.499857, y: -5.352017, z: 3.046877}
+  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_Children: []
+  m_Father: {fileID: 0}
+  m_RootOrder: 8
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1 &1932856556
 GameObject:
   m_ObjectHideFlags: 0
@@ -2535,9 +2581,6 @@ MonoBehaviour:
   starting_Pipe_Position: {x: 0, y: 0, z: 0}
   character: {fileID: 413841361}
   last_Pipe_End: horizontal_right
-  small_Straight_Pipe: {fileID: 1687824822}
-  small_90Deg_Turn_Pipe: {fileID: 1855687612}
-  S_Pipe: {fileID: 2027925504}
 --- !u!4 &1932856558
 Transform:
   m_ObjectHideFlags: 0
@@ -2762,7 +2805,7 @@ Transform:
   - {fileID: 1585638962}
   - {fileID: 1088205621}
   m_Father: {fileID: 0}
-  m_RootOrder: 12
+  m_RootOrder: 13
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!136 &2027925510
 CapsuleCollider:
